migrate to useResizeButton
